% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{big_family_size_ped}
\alias{big_family_size_ped}
\title{A large pedigree with big family sizes}
\format{
A data.table with 8 columns:
\describe{
  \item{Ind}{Individual ID}
  \item{Sire}{Sire ID}
  \item{Dam}{Dam ID}
  \item{Sex}{Sex of the individual}
  \item{Year}{Year of birth}
  \item{IndNum}{Numeric ID for individual}
  \item{SireNum}{Numeric ID for sire}
  \item{DamNum}{Numeric ID for dam}
}
}
\usage{
big_family_size_ped
}
\description{
A dataset containing a pedigree with many full-sib individuals per family.
}
\keyword{datasets}
