% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_graph.R
\name{compact_pedigree}
\alias{compact_pedigree}
\title{Compact pedigree by merging full siblings}
\usage{
compact_pedigree(ped_node, compact, h_ids)
}
\arguments{
\item{ped_node}{A data.table of nodes.}

\item{compact}{Logical, whether to compact.}

\item{h_ids}{Highlighted IDs to exempt from compaction.}
}
\description{
Compact pedigree by merging full siblings
}
\keyword{internal}
