% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complex_ped}
\alias{complex_ped}
\title{A complex pedigree}
\format{
A data.table with a standard pedigree structure.
}
\usage{
complex_ped
}
\description{
A dataset containing a large, complex pedigree covering about 100 generations,
useful for testing the performance and accuracy of partial inbreeding and similar calculations.
}
\keyword{datasets}
