% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_build_ainv_triplets}
\alias{cpp_build_ainv_triplets}
\title{Build A-Inverse Sparse Matrix Components (Henderson's Rules)}
\usage{
cpp_build_ainv_triplets(sire, dam, dii)
}
\description{
Build A-Inverse Sparse Matrix Components (Henderson's Rules)
}
