% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_calculate_A}
\alias{cpp_calculate_A}
\title{Calculate Additive Relationship Matrix (A) using Armadillo}
\usage{
cpp_calculate_A(sire, dam)
}
\description{
Calculate Additive Relationship Matrix (A) using Armadillo
}
