% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_calculate_inbreeding}
\alias{cpp_calculate_inbreeding}
\title{Calculate Inbreeding and Dii using Meuwissen & Luo (1992)}
\usage{
cpp_calculate_inbreeding(sire, dam)
}
\description{
Calculate Inbreeding and Dii using Meuwissen & Luo (1992)
}
