% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_invert_sympd}
\alias{cpp_invert_sympd}
\title{Invert Symmetric Positive-Definite Matrix (optimized)
Uses Cholesky decomposition, approximately 2x faster than general inversion
for symmetric positive-definite matrices}
\usage{
cpp_invert_sympd(M)
}
\description{
Invert Symmetric Positive-Definite Matrix (optimized)
Uses Cholesky decomposition, approximately 2x faster than general inversion
for symmetric positive-definite matrices
}
