% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deep_ped}
\alias{deep_ped}
\title{A deep pedigree}
\format{
A data.table with 4 columns:
\describe{
  \item{Ind}{Individual ID}
  \item{Sire}{Sire ID}
  \item{Dam}{Dam ID}
  \item{Sex}{Sex of the individual}
}
}
\usage{
deep_ped
}
\description{
A dataset containing a pedigree with many generations.
}
\keyword{datasets}
