% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedmatrix.R
\name{diag}
\alias{diag}
\title{Extract Matrix Diagonals}
\usage{
diag(x, ...)
}
\arguments{
\item{x}{A matrix, vector, or \code{pedmat} object.}

\item{...}{Additional arguments passed to \code{base::diag}.}
}
\value{
The diagonal vectors or a diagonal matrix.
}
\description{
Extract the diagonal of a pedmat object or other matrices.
This generic function extending \code{base::diag} to support \code{pedmat} objects.
}
