% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedexport.R
\name{.pedexport_char}
\alias{.pedexport_char}
\title{Build a character export table (ASReml format)}
\usage{
.pedexport_char(ped, missing = "0")
}
\arguments{
\item{ped}{A complete tidyped object.}

\item{missing}{Character scalar for missing parents.}
}
\value{
A data.table with columns animal, sire, dam.
}
\description{
Build a character export table (ASReml format)
}
\keyword{internal}
