% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedexport.R
\name{.pedexport_num}
\alias{.pedexport_num}
\title{Build a numeric (integer) export table}
\usage{
.pedexport_num(ped, missing = 0L)
}
\arguments{
\item{ped}{A complete tidyped object.}

\item{missing}{Integer scalar for missing parents.}
}
\value{
A data.table with columns IndNum, SireNum, DamNum.
}
\description{
Build a numeric (integer) export table
}
\keyword{internal}
