% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_style.R
\name{finalize_graph}
\alias{finalize_graph}
\title{Finalize graph and reindex IDs}
\usage{
finalize_graph(ped_node, ped_edge, h_ids, showf)
}
\description{
Finalize graph and reindex IDs
}
\keyword{internal}
