% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_graph.R
\name{generate_graph_structure}
\alias{generate_graph_structure}
\title{Generate edges and virtual family nodes}
\usage{
generate_graph_structure(ped_node, h_ids)
}
\arguments{
\item{ped_node}{A data.table of nodes.}

\item{h_ids}{Highlighted IDs.}
}
\description{
Generate edges and virtual family nodes
}
\keyword{internal}
