% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_style.R
\name{get_highlight_ids}
\alias{get_highlight_ids}
\title{Styling and finalizing pedigree graph}
\usage{
get_highlight_ids(ped, highlight, trace)
}
\description{
Styling and finalizing pedigree graph
}
\keyword{internal}
