% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{has_candidates}
\alias{has_candidates}
\title{Check whether a tidyped object contains candidate flags}
\usage{
has_candidates(x)
}
\arguments{
\item{x}{A tidyped object.}
}
\value{
Logical scalar.
}
\description{
Check whether a tidyped object contains candidate flags
}
