% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{has_inbreeding}
\alias{has_inbreeding}
\title{Check whether a tidyped object contains inbreeding coefficients}
\usage{
has_inbreeding(x)
}
\arguments{
\item{x}{A tidyped object.}
}
\value{
Logical scalar.
}
\description{
Check whether a tidyped object contains inbreeding coefficients
}
