% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inbred_ped}
\alias{inbred_ped}
\title{A highly inbred pedigree}
\format{
A data.table with 5 columns:
\describe{
  \item{Ind}{Individual ID}
  \item{Sire}{Sire ID}
  \item{Dam}{Dam ID}
  \item{Sex}{Sex of the individual}
  \item{Gen}{Generation number}
}
}
\usage{
inbred_ped
}
\description{
A simulated pedigree designed to demonstrate high levels of inbreeding and partial inbreeding decomposition.
Contains full-sib mating and backcrossing.
}
\keyword{datasets}
