% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbreed.R
\name{inbreed}
\alias{inbreed}
\title{Calculate inbreeding coefficients}
\usage{
inbreed(ped)
}
\arguments{
\item{ped}{A data.table including the pedigree tidied by the \code{\link{tidyped}} function.}
}
\value{
A data.table with an additional column \strong{f}.
}
\description{
\code{inbreed} function calculates the inbreeding coefficients for all individuals in a tidied pedigree.
}
\details{
This function takes a pedigree tidied by the \code{\link{tidyped}} function and calculates the inbreeding coefficients using the \code{makeDiiF} function from the \strong{nadiv} package. It prefers using numeric columns (\strong{IndNum}, \strong{SireNum}, \strong{DamNum}) if available, which is faster and more robust.
}
