% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_graph.R
\name{inject_missing_parents}
\alias{inject_missing_parents}
\title{Inject missing parents for subsetted pedigrees}
\usage{
inject_missing_parents(ped)
}
\arguments{
\item{ped}{A data.table containing pedigree info.}
}
\description{
Inject missing parents for subsetted pedigrees
}
\keyword{internal}
