% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{is_complete_pedigree}
\alias{is_complete_pedigree}
\title{Check whether all referenced parents are present}
\usage{
is_complete_pedigree(x)
}
\arguments{
\item{x}{A pedigree-like object.}
}
\value{
Logical scalar.
}
\description{
Check whether all referenced parents are present
}
\keyword{internal}
