% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{is_tidyped}
\alias{is_tidyped}
\title{Test if an object is a tidyped}
\usage{
is_tidyped(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical scalar.
}
\description{
Test if an object is a tidyped
}
