% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loop_ped}
\alias{loop_ped}
\title{A pedigree with loops}
\format{
A data.table with 3 columns:
\describe{
  \item{Ind}{Individual ID}
  \item{Sire}{Sire ID}
  \item{Dam}{Dam ID}
}
}
\usage{
loop_ped
}
\description{
A dataset containing a pedigree with circular mating loops.
}
\keyword{datasets}
