% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{new_tidyped}
\alias{new_tidyped}
\title{Internal constructor for tidyped class}
\usage{
new_tidyped(x)
}
\arguments{
\item{x}{A data.table object}
}
\value{
A tidyped object
}
\description{
Internal constructor for tidyped class
}
\keyword{internal}
