% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numped.R
\name{numped}
\alias{numped}
\title{Number a pedigree}
\usage{
numped(ped, ...)
}
\arguments{
\item{ped}{A data.table or data frame containing the pedigree, where the first three columns are \strong{Ind}, \strong{Sire}, and \strong{Dam}. Missing parents should be replaced with the default missing value \strong{NA}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A data.table containing the pedigree with new integer columns (\strong{IndNum}, \strong{SireNum}, and \strong{DamNum}).
}
\description{
The \code{numped} function generates integer IDs for the Ind, Sire, and Dam columns in the pedigree.
}
\details{
This function takes a pedigree and maps individual IDs to consecutive integers from 1 to n.
}
\keyword{internal}
