% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_graph.R
\name{ped2igraph}
\alias{ped2igraph}
\title{Convert pedigree to igraph structure}
\usage{
ped2igraph(
  ped,
  compact = FALSE,
  highlight = NULL,
  trace = FALSE,
  showf = FALSE,
  labelvar = NULL
)
}
\description{
Convert pedigree to igraph structure
}
\keyword{internal}
