% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedanalysis.R
\name{pedinbreed_class}
\alias{pedinbreed_class}
\title{Summarize Inbreeding Levels}
\usage{
pedinbreed_class(ped)
}
\arguments{
\item{ped}{A \code{tidyped} object.}
}
\value{
A \code{data.table} with counts and percentages for each F-group.
}
\description{
Classifies individuals into inbreeding levels based on their inbreeding 
coefficients (F) according to standard academic thresholds.
}
