% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{pedmeta}
\alias{pedmeta}
\title{Access pedigree metadata from a tidyped object}
\usage{
pedmeta(x)
}
\arguments{
\item{x}{A tidyped object.}
}
\value{
The \code{ped_meta} list, or \code{NULL} if not set.
}
\description{
Access pedigree metadata from a tidyped object
}
