% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{plot.tidyped}
\alias{plot.tidyped}
\title{Plot a tidy pedigree}
\usage{
\method{plot}{tidyped}(x, ...)
}
\arguments{
\item{x}{A \code{tidyped} object.}

\item{...}{Additional arguments passed to \code{\link{visped}}.}
}
\description{
Plot a tidy pedigree
}
