% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_graph.R
\name{prepare_initial_nodes}
\alias{prepare_initial_nodes}
\title{Prepare initial node table for igraph conversion}
\usage{
prepare_initial_nodes(ped, labelvar = NULL)
}
\arguments{
\item{ped}{A data.table containing pedigree info.}
}
\description{
Prepare initial node table for igraph conversion
}
\keyword{internal}
