% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_layout.R
\name{prepare_ped_graph}
\alias{prepare_ped_graph}
\title{Internal layout engine for pedigree visualization}
\usage{
prepare_ped_graph(
  ped,
  compact = FALSE,
  outline = FALSE,
  cex = NULL,
  highlight = NULL,
  trace = FALSE,
  showf = FALSE,
  labelvar = NULL,
  pagewidth = 200,
  symbolsize = 1,
  maxiter = 1000,
  ...
)
}
\description{
Internal layout engine for pedigree visualization
}
\keyword{internal}
