% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped.R
\name{prepare_visped_geninfo}
\alias{prepare_visped_geninfo}
\title{Attach display labels to the final plotted generation layers}
\usage{
prepare_visped_geninfo(gen_info, labels = NULL)
}
\description{
Attach display labels to the final plotted generation layers
}
\keyword{internal}
