% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped.R
\name{prepare_visped_genlab}
\alias{prepare_visped_genlab}
\title{Normalize generation-label display settings}
\usage{
prepare_visped_genlab(genlab)
}
\description{
Normalize generation-label display settings
}
\keyword{internal}
