% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped.R
\name{prepare_visped_labels}
\alias{prepare_visped_labels}
\title{Normalize custom labels before pedigree tidying}
\usage{
prepare_visped_labels(ped, labelvar)
}
\description{
Normalize custom labels before pedigree tidying
}
\keyword{internal}
