% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedanalysis.R
\name{print.pedcontrib}
\alias{print.pedcontrib}
\title{Print Founder and Ancestor Contributions}
\usage{
\method{print}{pedcontrib}(x, ...)
}
\arguments{
\item{x}{A \code{pedcontrib} object.}

\item{...}{Additional arguments.}
}
\description{
Print Founder and Ancestor Contributions
}
