% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedanalysis.R
\name{print.pediv}
\alias{print.pediv}
\title{Print Genetic Diversity Summary}
\usage{
\method{print}{pediv}(x, ...)
}
\arguments{
\item{x}{A \code{pediv} object.}

\item{...}{Additional arguments.}
}
\description{
Print Genetic Diversity Summary
}
