% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedanalysis.R
\name{print.pedstats}
\alias{print.pedstats}
\title{Print Pedigree Statistics}
\usage{
\method{print}{pedstats}(x, ...)
}
\arguments{
\item{x}{A \code{pedstats} object.}

\item{...}{Additional arguments.}
}
\description{
Print Pedigree Statistics
}
