% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{print.summary.tidyped}
\alias{print.summary.tidyped}
\title{Print method for summary.tidyped}
\usage{
\method{print}{summary.tidyped}(x, ...)
}
\arguments{
\item{x}{A summary.tidyped object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object, invisibly.
}
\description{
Print method for summary.tidyped
}
