% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{print.tidyped}
\alias{print.tidyped}
\title{Print method for tidyped pedigree}
\usage{
\method{print}{tidyped}(x, ...)
}
\arguments{
\item{x}{A tidyped object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for tidyped pedigree
}
