% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_layout.R
\name{repeloverlap}
\alias{repeloverlap}
\title{Repel overlapping nodes on the x-axis}
\usage{
repeloverlap(x)
}
\arguments{
\item{x}{A numeric vector of x positions.}
}
\value{
A numeric vector of unique x positions.
}
\description{
Repel overlapping nodes on the x-axis
}
\keyword{internal}
