% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_render.R
\name{resolve_visped_genlabcex}
\alias{resolve_visped_genlabcex}
\title{Resolve generation-label text size}
\usage{
resolve_visped_genlabcex(node_size, genlabcex = NULL)
}
\description{
Resolve generation-label text size
}
\keyword{internal}
