% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_ped}
\alias{simple_ped}
\title{A simple pedigree}
\format{
A data.table with 3 columns:
\describe{
  \item{Ind}{Individual ID}
  \item{Sire}{Sire ID}
  \item{Dam}{Dam ID}
}
}
\usage{
simple_ped
}
\description{
A small dataset containing a simple pedigree for demonstration.
}
\keyword{datasets}
