% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_ped}
\alias{small_ped}
\title{A small pedigree}
\format{
A data.frame with 3 columns:
\describe{
  \item{Ind}{Individual ID}
  \item{Sire}{Sire ID}
  \item{Dam}{Dam ID}
}
}
\usage{
small_ped
}
\description{
A small dataset containing a pedigree with some missing parents.
}
\keyword{datasets}
