% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortped.R
\name{sortped}
\alias{sortped}
\title{Sort a pedigree}
\usage{
sortped(ped, addgen = TRUE)
}
\arguments{
\item{ped}{A data.table or data frame containing the pedigree, where the first three columns are \strong{Ind}, \strong{Sire}, and \strong{Dam}. Missing parents should be replaced with the default missing value \strong{NA}.}

\item{addgen}{A logical value indicating whether the individual generation number will be inferred. Default is TRUE, which adds a \strong{Gen} column to the output.}
}
\value{
A data.table containing the sorted pedigree. Individual generations are inferred and a new column \strong{Gen} is added when \code{addgen = TRUE}. The \strong{Gen} column is an integer.
}
\description{
The \code{sortped} function assigns generations and sorts a pedigree.
}
\details{
This function takes a pedigree, assigns individuals to different generations, and sorts parents before offspring.
}
\keyword{internal}
