% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{validate_tidyped}
\alias{validate_tidyped}
\title{Internal validator for tidyped class}
\usage{
validate_tidyped(x)
}
\arguments{
\item{x}{A tidyped object}
}
\value{
The object if valid, otherwise an error
}
\description{
Internal validator for tidyped class
}
\keyword{internal}
