% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_style.R
\name{apply_node_styles}
\alias{apply_node_styles}
\title{Apply node styles (color, shape, highlighting)}
\usage{
apply_node_styles(ped_node, highlight_info)
}
\description{
Apply node styles (color, shape, highlighting)
}
\keyword{internal}
