% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped_render.R
\name{plot_ped_igraph}
\alias{plot_ped_igraph}
\title{Render pedigree graph using Two-Pass strategy}
\usage{
plot_ped_igraph(
  g,
  l,
  node_size,
  gen_info = NULL,
  genlab = FALSE,
  genlabcex = NULL,
  ...
)
}
\description{
Render pedigree graph using Two-Pass strategy
}
\keyword{internal}
