% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped.R
\name{plot_ped_igraph}
\alias{plot_ped_igraph}
\title{Internal plotting function for pedigree igraph}
\usage{
plot_ped_igraph(g, l, node_size, ...)
}
\arguments{
\item{g}{An igraph object.}

\item{l}{A layout matrix.}

\item{node_size}{Numeric node size.}

\item{...}{Additional arguments passed to \code{plot.igraph}.}
}
\description{
Internal plotting function for pedigree igraph
}
\keyword{internal}
