% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visped.R
\name{repeloverlap}
\alias{repeloverlap}
\title{Repel the overlapping nodes in x axis.}
\usage{
repeloverlap(x)
}
\arguments{
\item{x}{A vector including x positions with repeated values}
}
\value{
A vector including x positions with unique values
}
\description{
\code{repeloverlap} function converts repeated axis x positions to continuous positions.
}
\details{
This function takes a vector including axis x positions with repeated values, converts the repeated values of one x position to continuous positions. For example, x = c(1.2,2.1,2.1,2.1,3.2,4.6,5.7), where 2.1 is a axis x position with two repeated values; bl <- (3.2-2.1)/3; transformed x = c(1.2,2.1,2.1+bl,2.1+2*bl,3.2,4.6,5.7)
}
\keyword{internal}
